+++ src/dotnet/HackerRankSolutions/IProblem.cs
@@ -1,8 +1,0 @@
-﻿using HackerRankSolutions.ReaderWriter;
-
-namespace HackerRankSolutions;
-
-public interface IProblem
-{
-    void Solve(IReaderWriter readerWriter);
-}