+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 * Create new structured solution.
 * Merge all branches.
+* Move console project -> NUnit project.
+* Add Problem Solving (Basic) Certificate.
 
 ## 0.1 - 2021.07.21
 
 * Created repo.
+* Add some solutions.
+* Add solutions for 30-day-of-code tutorial